+++ src/optics_raytracer/pixelated_viewport.py
@@ -4,8 +4,8 @@ from .primitives import vector_dtype
 pixelated_viewport_dtype = np.dtype([
     ('middle_point', *vector_dtype),
     ('normal', *vector_dtype),
-    ('width', np.float64),
-    ('height', np.float64),
+    ('width', np.float32),
+    ('height', np.float32),
     ('u_vector', *vector_dtype),
     ('pixel_columns', np.int64),
     ('pixel_rows', np.int64),
