+++ src/optics_raytracer/primitives.py
@@ -1,5 +1,5 @@
 import numpy as np
 
-vector_dtype = (np.float64, (3,))
+vector_dtype = (np.float32, (3,))
 color_dtype = vector_dtype
 point_dtype = vector_dtype
